requirements.txt
# Core dependencies
openai>=1.10.0
pinecone-client>=3.0.0
langchain-community>=0.0.30
langchain-core>=0.1.35

# FastAPI and server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Utility and env management
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: for validation and request models
pydantic>=2.6.0

# Optional: for Google Sheets integration (if using)
gspread>=5.12.0
oauth2client>=4.1.3
